requirements.txt
# PDF processing
pdfplumber==0.10.3

# Text processing and utilities
python-dotenv==1.0.0
tqdm==4.66.1
tiktoken==0.5.2

# Data handling
pandas==2.1.4
pyarrow==14.0.1

# Logging and configuration
colorlog==6.8.0

# Azure OpenAI
openai==1.108.0

